[REFACTOR] config 파일 통합
